requirements.txt
pip>=21.3.1
PyYAML>=6.0
docker==5.0.3
GitPython==3.1.24